Use msgpack, implement versioned encryption
Blarg, we need to fix signing so that it actually signs the whole blob instead
of only signing specific parts of it. Not sure if this is actually any less
secure, but better to be safe than sorry!

This basically means that we need to use a unified meta/payload format for all
encrypted blobs, and that we need to serialize, mark, and sign the outer
container format ourselves, not using msgpack.

Ick. Ick ick ick. There's got to be a better way to serialize things than the
fixed-binary-ghetto format we were using before...

What's the penalty of just trying increasingly-older file format versions until
one works? As long as the user is alerted to the progress, hopefully not much...