Fix HMAC
We now encode and append the HMAC separately from everything else, in order to
strictly adhere to how HMAC works.

We use msgpack to encode all the information needed for decryption, then we
append a msgpack bin object _sans its body data_ (i.e. only its header data) to
represent the signature. We sign the blob _including this header_, then append
the signature itself, which will match the already-appended header.  We can do
this since the signature is always the same length, and this way we get a nice
all-msgpack stream of information, making decoding simple.

TODO: update all the tests!